# api/file/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Service API
    version: 0.0.1
paths:
    /resource/client/v1/file/sha:
        get:
            tags:
                - Service
            operationId: Service_GetFileBySha
            parameters:
                - name: sha
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/File'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /resource/client/v1/upload/prepare:
        post:
            tags:
                - Service
            operationId: Service_PrepareUploadFile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PrepareUploadFileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PrepareUploadFileReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /resource/v1/directories:
        get:
            tags:
                - Service
            operationId: Service_AllDirectory
            parameters:
                - name: parentId
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: app
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AllDirectoryReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /resource/v1/directory:
        put:
            tags:
                - Service
            operationId: Service_UpdateDirectory
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateDirectoryRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - Service
            operationId: Service_AddDirectory
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AddDirectoryRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Directory'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        delete:
            tags:
                - Service
            operationId: Service_DeleteDirectory
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: app
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /resource/v1/file:
        put:
            tags:
                - Service
            operationId: Service_UpdateFile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateFileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - Service
            operationId: Service_DeleteFile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DeleteFileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /resource/v1/file/sha:
        get:
            tags:
                - Service
            operationId: Service_GetFileBySha
            parameters:
                - name: sha
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/File'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /resource/v1/files:
        get:
            tags:
                - Service
            operationId: Service_PageFile
            parameters:
                - name: directoryId
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: app
                  in: query
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PageFileReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /resource/v1/upload/prepare:
        post:
            tags:
                - Service
            operationId: Service_PrepareUploadFile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PrepareUploadFileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PrepareUploadFileReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        AddDirectoryRequest:
            type: object
            properties:
                parentId:
                    type: integer
                    format: uint32
                name:
                    type: string
                app:
                    type: string
        AllDirectoryReply:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/Directory'
        DeleteFileRequest:
            type: object
            properties:
                ids:
                    type: array
                    items:
                        type: integer
                        format: uint32
                app:
                    type: string
                directoryId:
                    type: integer
                    format: uint32
        Directory:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                parentId:
                    type: integer
                    format: uint32
                app:
                    type: string
                name:
                    type: string
                createdAt:
                    type: integer
                    format: uint32
                updatedAt:
                    type: integer
                    format: uint32
        File:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                directoryId:
                    type: integer
                    format: uint32
                name:
                    type: string
                type:
                    type: string
                size:
                    type: integer
                    format: uint32
                sha:
                    type: string
                src:
                    type: string
                storage:
                    type: string
                createdAt:
                    type: integer
                    format: uint32
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        PageFileReply:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/File'
                total:
                    type: integer
                    format: uint32
        PrepareUploadFileReply:
            type: object
            properties:
                uploaded:
                    type: boolean
                src:
                    type: string
                chunkSize:
                    type: integer
                    format: uint32
                chunkCount:
                    type: integer
                    format: uint32
                uploadId:
                    type: string
                uploadChunks:
                    type: array
                    items:
                        type: integer
                        format: uint32
                sha:
                    type: string
        PrepareUploadFileRequest:
            type: object
            properties:
                directoryId:
                    type: integer
                    format: uint32
                directoryPath:
                    type: string
                app:
                    type: string
                name:
                    type: string
                sha:
                    type: string
                size:
                    type: integer
                    format: uint32
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        UpdateDirectoryRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                app:
                    type: string
                name:
                    type: string
        UpdateFileRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                app:
                    type: string
                name:
                    type: string
                directoryId:
                    type: integer
                    format: uint32
tags:
    - name: Service